refactor: extract PendingStopIcon and UserMessageBody from chat-message
