+++ src/components/features/chat/chat-message.tsx
@@ -7,141 +7,14 @@ import { StyledTooltip } from "#/components/shared/buttons/styled-tooltip";
 import { I18nKey } from "#/i18n/declaration";
 import { TextShimmer } from "#/components/shared/text-shimmer";
 import { MarkdownRenderer } from "../markdown/markdown-renderer";
+import { PendingStopIcon } from "./pending-stop-icon";
+import {
+  UserMessageBody,
+  chatBubbleMarkdownComponents,
+  USER_MESSAGE_LINE_HEIGHT_PX,
+} from "./user-message-body";
 
 export type ChatMessagePendingStatus = "sending" | "error";
-
-const USER_MESSAGE_MAX_LINES = 3;
-const USER_MESSAGE_LINE_HEIGHT_PX = 24;
-
-function PendingStopIcon({ className }: { className?: string }) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 24 24"
-      fill="none"
-      aria-hidden
-      className={className}
-    >
-      <circle
-        cx="12"
-        cy="12"
-        r="10"
-        className="fill-[var(--oh-foreground)] transition-colors duration-150 group-hover:fill-[var(--oh-text-secondary)]"
-      />
-      <rect
-        x="9"
-        y="9"
-        width="6"
-        height="6"
-        rx="1"
-        className="fill-[var(--oh-color-tertiary)]"
-      />
-    </svg>
-  );
-}
-
-const chatBubbleMarkdownComponents = {
-  p: ({ children }: React.ComponentProps<"p">) => (
-    <p className="m-0 leading-6">{children}</p>
-  ),
-};
-
-function UserMessageBody({
-  message,
-  isHovering,
-  isExpanded,
-  onTruncatableChange,
-}: {
-  message: string;
-  isHovering: boolean;
-  isExpanded: boolean;
-  onTruncatableChange: (truncatable: boolean) => void;
-}) {
-  const { t } = useTranslation("openhands");
-  const contentRef = React.useRef<HTMLDivElement>(null);
-  const [isTruncatable, setIsTruncatable] = React.useState(false);
-
-  React.useLayoutEffect(() => {
-    const content = contentRef.current;
-    if (!content || isExpanded) {
-      setIsTruncatable(false);
-      onTruncatableChange(false);
-      return undefined;
-    }
-
-    const measure = () => {
-      const lineHeight = Number.parseFloat(
-        getComputedStyle(content).lineHeight,
-      );
-      const linePx =
-        Number.isFinite(lineHeight) && lineHeight > 0
-          ? lineHeight
-          : USER_MESSAGE_LINE_HEIGHT_PX;
-      const maxHeight = USER_MESSAGE_MAX_LINES * linePx;
-      const newlineCount = (message.match(/\n/g) ?? []).length;
-
-      const truncatable =
-        content.scrollHeight > maxHeight + 1 ||
-        newlineCount >= USER_MESSAGE_MAX_LINES ||
-        message.trim().length > 220;
-
-      setIsTruncatable((previous) => {
-        if (previous !== truncatable) {
-          onTruncatableChange(truncatable);
-        }
-        return truncatable;
-      });
-    };
-
-    measure();
-
-    const observer = new ResizeObserver(measure);
-    observer.observe(content);
-
-    return () => observer.disconnect();
-  }, [message, isExpanded, onTruncatableChange]);
-
-  const isCollapsed = isTruncatable && !isExpanded;
-
-  return (
-    <div className="relative min-w-0">
-      <div
-        ref={contentRef}
-        className={cn(
-          "text-sm leading-6 whitespace-normal [word-break:break-word]",
-          isCollapsed && "line-clamp-3",
-        )}
-      >
-        <MarkdownRenderer
-          includeStandard
-          includeHeadings
-          components={chatBubbleMarkdownComponents}
-        >
-          {message}
-        </MarkdownRenderer>
-      </div>
-
-      {isCollapsed ? (
-        <>
-          <div
-            aria-hidden
-            data-testid="chat-message-truncation-gradient"
-            className="pointer-events-none absolute inset-x-0 bottom-0 h-10 bg-gradient-to-t from-tertiary to-transparent"
-          />
-          <span
-            data-testid="chat-message-view-more"
-            className={cn(
-              "pointer-events-none absolute bottom-1 left-1/2 z-10 inline-flex -translate-x-1/2 items-center rounded-full border border-[var(--oh-border-subtle)] bg-[var(--oh-surface-raised)] px-2.5 py-0.5 text-xs font-normal text-[var(--oh-foreground)] transition-opacity duration-150",
-              isHovering ? "opacity-100" : "opacity-0",
-            )}
-          >
-            {t(I18nKey.COMMON$VIEW_MORE)}
-          </span>
-        </>
-      ) : null}
-    </div>
-  );
-}
 
 interface ChatMessageProps {
   type: SourceType;
